refactor: clear clippy lib warnings (cosmetic, semantics-preserving)
Machine-applicable clippy fixes across push_ops/result_ops/runtime/script_builder
/splice (collapsible ifs, redundant cast/ref, manual is_multiple_of), plus an
explicit allow(clippy::wrong_self_convention) on the Interoperable trait whose
`from_*` methods deliberately take &mut self (they deserialize INTO self,
mirroring C# IInteroperable.FromStackItem/FromReplica). Full suite (180) green =
no semantic change; `cargo clippy --lib` now 0 warnings.

+++ src/interpreter/executor/push_ops.rs
@@ -181,7 +181,7 @@ pub(super) fn execute(
             if target < 0 || target as usize > script.len() {
                 return Err("PUSHA target out of bounds".to_string());
             }
-            stack.push(StackValue::Pointer(target as i64));
+            stack.push(StackValue::Pointer(target));
             ip += 5;
             finish!(Dispatch::Continue);
         }
